Move TraceHeaderSpec to separate file
