+++ segfast/__init__.py
@@ -2,5 +2,6 @@
 from .segyio_loader import SegyioLoader, SafeSegyioLoader
 from .memmap_loader import MemmapLoader
 from .loader import Loader, File, open #pylint: disable=redefined-builtin
+from .trace_header_spec import TraceHeaderSpec
 
 __version__ = '1.0.1'
